:truck: refactor for better organisation
